requirements.txt
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.5
html2text==2020.1.16
feedparser==6.0.10
pytz==2023.3